Prevent mod_md segfault with non letsencrypt acme
We pull in these commits from upstream that sync up mod_md:
c0cff48bccf6607333a6841f723660c0083929f4
7da5270ffe75101b19e1ca31b287c5afa2d44e8a

Both of these patches should be able to be
dropped once 2.4.48 is released.

We also need to run autoconf, as configure is not shipped in the trunk
repo, we need to regen it to get build mod_md after the changes.