postgresql use ISM by default and add DISM support.
This adds support for using Intimate Shared Memory using the sysv SHM
methods and Dynamic Intimate Shared Memory when available by checking
for the SHM_PAGEABLE flag in <sys/shm.h>. This makes it possible to have
ISM/DISM without needing to set the shared_memory_type and
dynamic_shared_memory_type to sysv in postgresql.conf. The setting of
shared_memory_type is something that seems to origin from Joyent for
Manta and is available in upstream versions 12 and later so we only
patch those versions for now. Changed --with-xslt to --with-libxslt so
the option is recognized by configure. For version 14 added the new lz4
support and llvm for JIT compiling as that is stable these days and no
longer highly experimental.